+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changes
+
+* The `Plugin.__call__` method now allows arbitrary keyword arguments to be
+  passed on to the `Plugin.run` method
+* The `Database` class now acts like a sequence
+
 ## [0.2.0]
 
 ### Added